chore: migration fixes Refs: PL-292
